# Compiling several regexes ahead of time using re2go for better performance