--- a/go.mod
+++ b/go.mod
@@ -4,10 +4,10 @@ go 1.20
 
 require (
 	github.com/araddon/dateparse v0.0.0-20210429162001-6b43995a97de
-	github.com/go-shiori/dom v0.0.0-20210627111528-4e4722cd0d65
+	github.com/go-shiori/dom v0.0.0-20230515143342-73569d674e1c
 	github.com/sergi/go-diff v1.1.0
 	github.com/spf13/cobra v1.0.0
-	golang.org/x/net v0.9.0
+	golang.org/x/net v0.29.0
 )
 
 require (
@@ -15,5 +15,5 @@ require (
 	github.com/gogs/chardet v0.0.0-20211120154057-b7413eaefb8f // indirect
 	github.com/inconshreveable/mousetrap v1.0.0 // indirect
 	github.com/spf13/pflag v1.0.5 // indirect
-	golang.org/x/text v0.9.0 // indirect
+	golang.org/x/text v0.18.0 // indirect
 )