--- a/parser.go
+++ b/parser.go
@@ -24,7 +24,6 @@ var (
 	rxPositive             = regexp.MustCompile(`(?i)article|body|content|entry|hentry|h-entry|main|page|pagination|post|text|blog|story`)
 	rxNegative             = regexp.MustCompile(`(?i)-ad-|hidden|^hid$| hid$| hid |^hid |banner|combx|comment|com-|contact|foot|footer|footnote|gdpr|masthead|media|meta|outbrain|promo|related|scroll|share|shoutbox|sidebar|skyscraper|sponsor|shopping|tags|tool|widget`)
 	rxByline               = regexp.MustCompile(`(?i)byline|author|dateline|writtenby|p-author`)
-	rxNormalize            = regexp.MustCompile(`(?i)\s{2,}`)
 	rxVideosx              = regexp.MustCompile(`(?i)//(www\.)?((dailymotion|youtube|youtube-nocookie|player\.vimeo|v\.qq)\.com|(archive|upload\.wikimedia)\.org|player\.twitch\.tv)`)
 	rxTokenize             = regexp.MustCompile(`(?i)\W+`)
 	rxWhitespace           = regexp.MustCompile(`(?i)^\s*$`)
@@ -410,7 +409,7 @@ func (ps *Parser) getArticleTitle() string {
 	}
 
 	curTitle = strings.TrimSpace(curTitle)
-	curTitle = rxNormalize.ReplaceAllString(curTitle, " ")
+	curTitle = re2go.NormalizeSpaces(curTitle)
 	// If we now have 4 words or fewer as our title, and either no
 	// 'hierarchical' separators (\, /, > or ») were found in the original
 	// title or we decreased the number of words by more than 1 word, use
@@ -1668,7 +1667,7 @@ func (ps *Parser) isWhitespace(node *html.Node) bool {
 func (ps *Parser) getInnerText(node *html.Node, normalizeSpaces bool) string {
 	textContent := strings.TrimSpace(dom.TextContent(node))
 	if normalizeSpaces {
-		textContent = rxNormalize.ReplaceAllString(textContent, " ")
+		textContent = re2go.NormalizeSpaces(textContent)
 	}
 	return textContent
 }