… go.sum
… internal/re2go/base.go
… internal/re2go/check-byline.go
… internal/re2go/class-weight.go
… internal/re2go/class-weight.re
… internal/re2go/grab-article.go
… internal/re2go/normalize.go
… internal/re2go/normalize.re