## What is it?

I was running benchmark comparing several content extractors. According to my benchmark go-readability able to extract 983 web pages in 10.64 seconds, which is pretty fast (only 10ms per page). However I was wondering if we could make it faster, so I used the profiler to look for parts to improve.

Here is the result. As you can see, go-readability took 10.64 seconds to finish the benchmark:

![The old go-readability performance](https://github.com/user-attachments/assets/95fef3d5-0951-44e7-9e9a-2b465d17b63c)

From that graph above, in a glance we can see that regular expressions took most of the time, so I filter it to get the exact number:

![How much regex we use in old version](https://github.com/user-attachments/assets/42cd19d2-10d1-4e3b-9576-1f54bbacf48b)

So, from the entire 10.64 seconds, regular expressions alone use 7.48 seconds or around 70% of the entire process.

I was wondering how to improve the speed since Go' regex is known to be [slow](https://github.com/golang/go/issues/26623#issuecomment-1899965448). The first alternative is to use regex package that uses cgo like [`github.com/wasilibs/go-re2`](https://github.com/wasilibs/go-re2), however cgo make it hard to cross compile Go' application.

After looking around, I found about [re2go](https://re2c.org/manual/manual_go.html#regular-expressions). It's a compiler that compile regular expressions ahead of time into Go' code. It's pretty interesting:

- It uses a new-ish lookahead TDFA algorithm, which supposed to be faster than Go' regex that use NFA and one-pass algorithm.
- It has a long history and is still actively maintained.
- It's already used in production by various open source projects like PHP and Ninja.
- After the initial learning curve, it's actually pretty easy to use.
- From [my regex benchmark](https://github.com/RadhiFadlillah/go-regex-benchmark-web/tree/concurrent), for some case it's actually faster than `wasilibs/go-re2` that uses cgo binding.

## Which regex to replace?

Obviously we can't replace every regex with re2go, so I looked at the profile to see which regex to replace. I found several regex that take most of the time:

1. First are `rxUnlikelyCandidates`, `rxOkMaybeItsACandidate` and `rxCommas` that used in `grabArticle` function. All together, those 3 regex take around 1.5 seconds.
2. Next is `rxNormalize` which used to remove double spaces in function `getInnerText`. This one regex alone take around 4.1 seconds.
3. Next is `rxByline` that used in function `checkByline`. This one take around 0.6 seconds.
4. Finally are `rxNegative` and `rxPositive` which used by `getClassWeight` function. Together those two regex take around 0.9 seconds.

All together take around 7.1 seconds, which is around 95% of the entire regex processes.

## The result

Here is the new profile after converting those regex with re2go:

![The result](https://github.com/user-attachments/assets/7749b22d-1910-4aff-9f13-371ddd95c6d8)

Now go-readability able to extract 983 web pages in 4.02 seconds, or 2.5x faster than the old one. As you can see now there are no regex in sight. The best thing is it's still pure Go code without cgo binding, so it's still easy to cross compile.

For the cons, there are additional burdens in us the maintainer since we need to convert the basic regex to re2go templates. However I believe it's worth it since the performance is increased that much.

## Additional notes

Besides converting those regex I also added tests for the code generated by re2go. I also updated all dependencies to the latest one.

If you want to see the profiles yourself, you can download it here: [readability-profile.zip](https://github.com/user-attachments/files/17037558/readability-profile.zip)